requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==41.0.0
cssselect==1.2.0
decorator==5.1.1
executing==1.2.0
Faker==18.10.1
filelock==3.12.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ingredient-parser-nlp==0.1.0b1
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
jmespath==1.0.1
joblib==1.2.0
lxml==4.9.2
matplotlib-inline==0.1.6
nltk==3.8.1
packaging==23.1
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
Protego==0.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.15.1
pyOpenSSL==23.2.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
queuelib==1.6.2
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
scrapeops-scrapy-proxy-sdk==1.0
Scrapy==2.9.0
service-identity==21.1.0
six==1.16.0
stack-data==0.6.2
text-unidecode==1.3
tldextract==3.4.4
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
typing_extensions==4.6.2
urllib3==2.0.2
w3lib==2.1.1
wcwidth==0.2.6
zope.interface==6.0
